Issue29: Moved Controller.js under framework folder and updated the imports accordingly
