+++ _templates/router/new/gen.ejs.t
@@ -4,7 +4,7 @@ to: src/routes/<%= h.inflection.pluralize(name) %>.js
 
 const {Router} = require('express');
 const HasScopes = require('../app/http/middleware/HasScopes');
-const { <%= h.inflection.pluralize(Name) %>Controller } = require('../app/http/controllers/<%= h.inflection.pluralize(Name) %>Controller');
+const { <%= h.inflection.pluralize(Name) %>Controller } = require('../framework/controllers/<%= h.inflection.pluralize(Name) %>Controller');
 const { Create<%= h.inflection.singularize(Name) %>Request: CreateRequest } = require('../app/http/requests/<%= h.inflection.pluralize(name) %>/Create<%= h.inflection.singularize(Name) %>Request');
 const { List<%= h.inflection.pluralize(Name) %>Request: ListRequest } = require('../app/http/requests/<%= h.inflection.pluralize(name) %>/List<%= h.inflection.pluralize(Name) %>Request');
 const { Retrieve<%= h.inflection.singularize(Name) %>Request: GetRequest } = require('../app/http/requests/<%= h.inflection.pluralize(name) %>/Retrieve<%= h.inflection.singularize(Name) %>Request');
