+++ src/app/http/controllers/CropsController.js
@@ -1,5 +1,5 @@
 const { Crop } = require('../../models/Crop');
-const { Controller } = require('./Controller');
+const { Controller } = require('../../../framework/controllers/Controller');
 const { PaginatedCollection } = require('../resources/PaginatedCollection');
 const { Resource } = require('../resources/Resource');
 const { CreatedResource } = require('../resources/CreatedResource');
