+++ src/app/http/controllers/GroupsController.js
@@ -1,5 +1,5 @@
 const { Group } = require('../../models/Group');
-const { Controller } = require('./Controller');
+const { Controller } = require('../../../framework/controllers/Controller');
 const { PaginatedCollection } = require('../resources/PaginatedCollection');
 const { Resource } = require('../resources/Resource');
 const { CreatedResource } = require('../resources/CreatedResource');
